+++ app/src/main/java/com/realityexpander/tasky/auth_feature/presentation/splash_screen/MainActivityViewModel.kt
@@ -83,6 +83,7 @@ class MainActivityViewModel @Inject constructor(
     }
 }
 
+// NOTE: LEAVE FOR REFERENCE
 // to scale the splash screen if XML/SVG:
 // put <path> inside a group as so:
 // <group
