+++ app/src/main/java/com/realityexpander/tasky/di/WorkManagerInitializer.kt
@@ -1,7 +1,6 @@
 package com.realityexpander.tasky.di
 
 import android.content.Context
-import android.util.Log
 import androidx.startup.Initializer
 import androidx.work.Configuration
 import androidx.work.WorkManager
@@ -10,6 +9,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import logcat.logcat
 import javax.inject.Singleton
 
 @Module
@@ -21,7 +21,8 @@ object WorkManagerInitializer : Initializer<WorkManager> {
     override fun create(@ApplicationContext context: Context): WorkManager {
         val configuration = Configuration.Builder().build()
         WorkManager.initialize(context, configuration)
-        Log.d("Hilt Init", "WorkManager initialized by Hilt")
+
+        logcat {"WorkManager initialized by Hilt" }
         return WorkManager.getInstance(context)
     }
